Refactor: readability improvements across Rust backend and React frontend
Split the 1,315-line availability_service.rs into focused submodules:
- checker.rs: product checking and result processing
- comparison.rs: price comparison and stock transition detection
- summary.rs: bulk check summary building
- types.rs: shared data types

Replace boolean flags with StringContext enum in GTM dataLayer parser
to make impossible states unrepresentable. Add comprehensive module-level
documentation for scraper strategy chain and price comparison logic.

Consolidate 7 useMemo calls into 3 logical groups in usePriceFormatting
hook. Fix pre-existing test compilation issues in domain_setting_service
and scraper tests.

Closes #42
