+++ apps/desktop/src-tauri/crates/domain/src/services/availability/checker.rs
@@ -1,132 +1,24 @@
-//! Availability service for checking product availability.
+//! Product availability checking and result processing.
 
 use sea_orm::DatabaseConnection;
-use serde::Serialize;
 use uuid::Uuid;
 
 use crate::entities::availability_check::AvailabilityStatus;
 use crate::entities::prelude::{AvailabilityCheckModel, ProductModel};
 use crate::repositories::{AvailabilityCheckRepository, CreateCheckParams, ProductRepository};
+use crate::services::scraper::has_path_locale;
+use crate::services::{NotificationService, ScraperService};
 use product_stalker_core::AppError;
 
-use product_stalker_core::services::notification_helpers::NotificationData;
-
-use super::scraper::has_path_locale;
-use super::{currency, NotificationService, ScraperService};
-
-/// Result of a single product availability check in a bulk operation
-#[derive(Debug, Clone, Default, Serialize)]
-pub struct BulkCheckResult {
-    pub product_id: String,
-    pub product_name: String,
-    pub status: AvailabilityStatus,
-    pub previous_status: Option<AvailabilityStatus>,
-    pub is_back_in_stock: bool,
-    pub price_minor_units: Option<i64>,
-    pub price_currency: Option<String>,
-    pub currency_exponent: Option<u32>,
-    pub today_average_price_minor_units: Option<i64>,
-    pub yesterday_average_price_minor_units: Option<i64>,
-    pub is_price_drop: bool,
-    pub error: Option<String>,
-}
-
-/// Summary of a bulk check operation
-#[derive(Debug, Clone, Serialize)]
-pub struct BulkCheckSummary {
-    pub total: usize,
-    pub successful: usize,
-    pub failed: usize,
-    pub back_in_stock_count: usize,
-    pub price_drop_count: usize,
-    pub results: Vec<BulkCheckResult>,
-}
-
-/// Result of an availability check with optional notification data
-#[derive(Debug, Serialize)]
-pub struct CheckResultWithNotification {
-    pub check: AvailabilityCheckModel,
-    pub notification: Option<NotificationData>,
-    pub daily_comparison: DailyPriceComparison,
-}
-
-/// Result of processing a single availability check
-pub struct CheckProcessingResult {
-    pub status: AvailabilityStatus,
-    pub price_minor_units: Option<i64>,
-    pub price_currency: Option<String>,
-    pub error: Option<String>,
-    pub is_back_in_stock: bool,
-    pub is_price_drop: bool,
-}
-
-/// Context for checking a single product in a bulk operation
-pub struct ProductCheckContext {
-    pub previous_status: Option<AvailabilityStatus>,
-}
-
-/// Accumulated counters for bulk check results
-#[derive(Default)]
-pub struct BulkCheckCounters {
-    pub successful: usize,
-    pub failed: usize,
-    pub back_in_stock_count: usize,
-    pub price_drop_count: usize,
-}
-
-/// Result of comparing today's average price vs yesterday's average price
-#[derive(Debug, Clone, Default, Serialize)]
-pub struct DailyPriceComparison {
-    pub today_average_minor_units: Option<i64>,
-    pub yesterday_average_minor_units: Option<i64>,
-}
-
-impl BulkCheckResult {
-    /// Build a result from a successful processing result with daily comparison data
-    pub fn from_processing_result(
-        product: &ProductModel,
-        result: &CheckProcessingResult,
-        context: &ProductCheckContext,
-        daily_comparison: &DailyPriceComparison,
-    ) -> Self {
-        let currency_exponent = result
-            .price_currency
-            .as_deref()
-            .or(product.currency.as_deref())
-            .map(currency::currency_exponent);
-        Self {
-            product_id: product.id.to_string(),
-            product_name: product.name.clone(),
-            status: result.status.clone(),
-            previous_status: context.previous_status.clone(),
-            is_back_in_stock: result.is_back_in_stock,
-            price_minor_units: result.price_minor_units,
-            price_currency: result.price_currency.clone(),
-            currency_exponent,
-            today_average_price_minor_units: daily_comparison.today_average_minor_units,
-            yesterday_average_price_minor_units: daily_comparison.yesterday_average_minor_units,
-            is_price_drop: result.is_price_drop,
-            error: result.error.clone(),
-        }
-    }
-
-    /// Build an error result when context or infrastructure fails
-    pub fn error_for_product(product: &ProductModel, error_message: String) -> Self {
-        Self {
-            product_id: product.id.to_string(),
-            product_name: product.name.clone(),
-            error: Some(error_message),
-            ..Default::default()
-        }
-    }
-}
-
-/// Service layer for availability checking business logic
-pub struct AvailabilityService;
+use super::types::{
+    BulkCheckResult, CheckProcessingResult, CheckResultWithNotification, DailyPriceComparison,
+    ProductCheckContext,
+};
+use super::AvailabilityService;
 
 impl AvailabilityService {
     /// Build CreateCheckParams from a successful scraping result
-    fn params_from_success(result: super::scraper::ScrapingResult) -> CreateCheckParams {
+    fn params_from_success(result: crate::services::scraper::ScrapingResult) -> CreateCheckParams {
         CreateCheckParams {
             status: result.status,
             raw_availability: result.raw_availability,
@@ -371,22 +263,6 @@ impl AvailabilityService {
         (bulk_result, result)
     }
 
-    /// Build summary from collected results
-    pub fn build_summary_from_results(
-        total: usize,
-        results: Vec<(BulkCheckResult, CheckProcessingResult)>,
-    ) -> BulkCheckSummary {
-        let mut counters = BulkCheckCounters::default();
-        let mut bulk_results = Vec::with_capacity(results.len());
-
-        for (bulk_result, processing_result) in results {
-            Self::update_counters(&mut counters, &processing_result);
-            bulk_results.push(bulk_result);
-        }
-
-        Self::build_bulk_summary(total, counters, bulk_results)
-    }
-
     /// Get the context needed before checking a product (previous status)
     pub async fn get_product_check_context(
         conn: &DatabaseConnection,
@@ -396,100 +272,6 @@ impl AvailabilityService {
         let previous_status = previous_check.as_ref().map(|c| c.status_enum());
 
         Ok(ProductCheckContext { previous_status })
-    }
-
-    /// Update counters based on the check result
-    pub fn update_counters(counters: &mut BulkCheckCounters, result: &CheckProcessingResult) {
-        if result.error.is_some() {
-            counters.failed += 1;
-            return;
-        }
-
-        counters.successful += 1;
-        if result.is_back_in_stock {
-            counters.back_in_stock_count += 1;
-        }
-        if result.is_price_drop {
-            counters.price_drop_count += 1;
-        }
-    }
-
-    /// Build the final bulk check summary from counters and results
-    pub fn build_bulk_summary(
-        total: usize,
-        counters: BulkCheckCounters,
-        results: Vec<BulkCheckResult>,
-    ) -> BulkCheckSummary {
-        BulkCheckSummary {
-            total,
-            successful: counters.successful,
-            failed: counters.failed,
-            back_in_stock_count: counters.back_in_stock_count,
-            price_drop_count: counters.price_drop_count,
-            results,
-        }
-    }
-
-    /// Determines if a product has transitioned back to being in stock.
-    ///
-    /// A product is considered "back in stock" only if:
-    /// 1. There was a previous check (first check doesn't count as "back")
-    /// 2. The previous status was NOT in_stock
-    /// 3. The new status IS in_stock
-    ///
-    /// This ensures we only notify users about meaningful transitions,
-    /// not products that were always in stock or are being checked for the first time.
-    pub fn is_back_in_stock(
-        previous_status: &Option<AvailabilityStatus>,
-        new_status: &AvailabilityStatus,
-    ) -> bool {
-        match previous_status {
-            Some(prev) => {
-                *prev != AvailabilityStatus::InStock && *new_status == AvailabilityStatus::InStock
-            }
-            None => false,
-        }
-    }
-
-    /// Check if today's average price dropped compared to yesterday's
-    pub fn is_price_drop(yesterday_average: Option<i64>, today_average: Option<i64>) -> bool {
-        match (yesterday_average, today_average) {
-            (Some(prev), Some(new)) => new < prev,
-            _ => false, // No price drop if either is None
-        }
-    }
-
-    /// Get today's and yesterday's average prices for comparison
-    ///
-    /// Uses rolling 24-hour windows instead of calendar dates for timezone
-    /// resilience. "Today" = last 24 hours, "Yesterday" = 24–48 hours ago.
-    pub async fn get_daily_price_comparison(
-        conn: &DatabaseConnection,
-        product_id: Uuid,
-    ) -> Result<DailyPriceComparison, AppError> {
-        let now = chrono::Utc::now();
-        let twenty_four_hours_ago = now - chrono::Duration::hours(24);
-        let forty_eight_hours_ago = now - chrono::Duration::hours(48);
-
-        let today_average = AvailabilityCheckRepository::get_average_price_for_period(
-            conn,
-            product_id,
-            twenty_four_hours_ago,
-            now,
-        )
-        .await?;
-        let yesterday_average = AvailabilityCheckRepository::get_average_price_for_period(
-            conn,
-            product_id,
-            forty_eight_hours_ago,
-            twenty_four_hours_ago,
-        )
-        .await?;
-
-        Ok(DailyPriceComparison {
-            today_average_minor_units: today_average,
-            yesterday_average_minor_units: yesterday_average,
-        })
     }
 
     /// Get the latest availability check for a product
@@ -558,21 +340,6 @@ impl AvailabilityService {
             notification,
             daily_comparison,
         })
-    }
-
-    /// Build notification data for a bulk check using pre-fetched settings
-    ///
-    /// Delegates to NotificationService for actual notification composition.
-    pub fn build_bulk_notification_with_settings(
-        enable_notifications: bool,
-        summary: &BulkCheckSummary,
-    ) -> Option<NotificationData> {
-        NotificationService::build_bulk_notification(
-            enable_notifications,
-            summary.back_in_stock_count,
-            summary.price_drop_count,
-            &summary.results,
-        )
     }
 }
 
@@ -703,398 +470,10 @@ mod tests {
             let conn = setup_availability_db().await;
             let fake_id = Uuid::new_v4();
 
-            let result = AvailabilityService::check_product(&conn, fake_id, false).await;
+            let result = AvailabilityService::check_product(&conn, fake_id, false, false, 14).await;
 
             assert!(result.is_err());
             assert!(matches!(result, Err(AppError::NotFound(_))));
-        }
-    }
-
-    /// Tests for is_back_in_stock logic
-    mod back_in_stock_tests {
-        use super::*;
-
-        #[test]
-        fn test_from_out_of_stock() {
-            let previous = Some(AvailabilityStatus::OutOfStock);
-            assert!(AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::InStock
-            ));
-        }
-
-        #[test]
-        fn test_from_back_order() {
-            let previous = Some(AvailabilityStatus::BackOrder);
-            assert!(AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::InStock
-            ));
-        }
-
-        #[test]
-        fn test_from_unknown() {
-            let previous = Some(AvailabilityStatus::Unknown);
-            assert!(AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::InStock
-            ));
-        }
-
-        #[test]
-        fn test_already_in_stock() {
-            let previous = Some(AvailabilityStatus::InStock);
-            assert!(!AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::InStock
-            ));
-        }
-
-        #[test]
-        fn test_still_out_of_stock() {
-            let previous = Some(AvailabilityStatus::OutOfStock);
-            assert!(!AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::OutOfStock
-            ));
-        }
-
-        #[test]
-        fn test_no_previous() {
-            let previous: Option<AvailabilityStatus> = None;
-            assert!(!AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::InStock
-            ));
-        }
-
-        #[test]
-        fn test_to_out_of_stock() {
-            let previous = Some(AvailabilityStatus::InStock);
-            assert!(!AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::OutOfStock
-            ));
-        }
-
-        #[test]
-        fn test_to_back_order() {
-            let previous = Some(AvailabilityStatus::InStock);
-            assert!(!AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::BackOrder
-            ));
-        }
-
-        #[test]
-        fn test_to_unknown() {
-            let previous = Some(AvailabilityStatus::InStock);
-            assert!(!AvailabilityService::is_back_in_stock(
-                &previous,
-                &AvailabilityStatus::Unknown
-            ));
-        }
-    }
-
-    /// Tests for is_price_drop logic
-    mod price_drop_tests {
-        use super::*;
-
-        #[test]
-        fn test_from_higher() {
-            assert!(AvailabilityService::is_price_drop(Some(10000), Some(8000)));
-        }
-
-        #[test]
-        fn test_same_price() {
-            assert!(!AvailabilityService::is_price_drop(
-                Some(10000),
-                Some(10000)
-            ));
-        }
-
-        #[test]
-        fn test_price_increase() {
-            assert!(!AvailabilityService::is_price_drop(Some(8000), Some(10000)));
-        }
-
-        #[test]
-        fn test_no_previous() {
-            assert!(!AvailabilityService::is_price_drop(None, Some(10000)));
-        }
-
-        #[test]
-        fn test_no_new() {
-            assert!(!AvailabilityService::is_price_drop(Some(10000), None));
-        }
-
-        #[test]
-        fn test_both_none() {
-            assert!(!AvailabilityService::is_price_drop(None, None));
-        }
-    }
-
-    /// Tests for BulkCheckResult struct
-    mod bulk_check_result_tests {
-        use super::*;
-
-        #[test]
-        fn test_serialize() {
-            let result = BulkCheckResult {
-                product_id: "test-id-123".to_string(),
-                product_name: "Test Product".to_string(),
-                status: AvailabilityStatus::InStock,
-                previous_status: Some(AvailabilityStatus::OutOfStock),
-                is_back_in_stock: true,
-                price_minor_units: Some(78900),
-                price_currency: Some("USD".to_string()),
-                today_average_price_minor_units: Some(78900),
-                yesterday_average_price_minor_units: Some(89900),
-                is_price_drop: true,
-                ..Default::default()
-            };
-            let json = serde_json::to_string(&result).unwrap();
-            assert!(json.contains("test-id-123"));
-            assert!(json.contains("Test Product"));
-            assert!(json.contains("in_stock"));
-            assert!(json.contains("out_of_stock"));
-            assert!(json.contains("78900"));
-        }
-
-        #[test]
-        fn test_with_error() {
-            let result = BulkCheckResult {
-                product_id: "error-id".to_string(),
-                product_name: "Error Product".to_string(),
-                error: Some("Failed to fetch".to_string()),
-                ..Default::default()
-            };
-            let json = serde_json::to_string(&result).unwrap();
-            assert!(json.contains("Failed to fetch"));
-            assert!(json.contains("unknown"));
-        }
-    }
-
-    /// Tests for BulkCheckSummary struct
-    mod bulk_check_summary_tests {
-        use super::*;
-
-        #[test]
-        fn test_serialize() {
-            let summary = BulkCheckSummary {
-                total: 10,
-                successful: 8,
-                failed: 2,
-                back_in_stock_count: 3,
-                price_drop_count: 2,
-                results: vec![],
-            };
-            let json = serde_json::to_string(&summary).unwrap();
-            assert!(json.contains("10"));
-            assert!(json.contains("\"successful\":8"));
-            assert!(json.contains("\"failed\":2"));
-            assert!(json.contains("\"back_in_stock_count\":3"));
-            assert!(json.contains("\"price_drop_count\":2"));
-        }
-
-        #[test]
-        fn test_with_results() {
-            let result = BulkCheckResult {
-                product_id: "p1".to_string(),
-                product_name: "Product 1".to_string(),
-                status: AvailabilityStatus::InStock,
-                previous_status: Some(AvailabilityStatus::OutOfStock),
-                is_back_in_stock: true,
-                price_minor_units: Some(78900),
-                price_currency: Some("USD".to_string()),
-                ..Default::default()
-            };
-            let summary = BulkCheckSummary {
-                total: 1,
-                successful: 1,
-                failed: 0,
-                back_in_stock_count: 1,
-                price_drop_count: 0,
-                results: vec![result],
-            };
-            let json = serde_json::to_string(&summary).unwrap();
-            assert!(json.contains("Product 1"));
-            assert!(json.contains("p1"));
-        }
-    }
-
-    /// Tests for CheckResultWithNotification struct
-    mod check_result_with_notification_tests {
-        use super::*;
-
-        #[test]
-        fn test_serialize() {
-            let check = AvailabilityCheckModel {
-                id: Uuid::new_v4(),
-                product_id: Uuid::new_v4(),
-                status: "in_stock".to_string(),
-                raw_availability: Some("http://schema.org/InStock".to_string()),
-                error_message: None,
-                checked_at: chrono::Utc::now(),
-                price_minor_units: Some(78900),
-                price_currency: Some("USD".to_string()),
-                raw_price: Some("789.00".to_string()),
-            };
-            let result = CheckResultWithNotification {
-                check,
-                notification: Some(NotificationData {
-                    title: "Back in Stock!".to_string(),
-                    body: "Product available".to_string(),
-                }),
-                daily_comparison: DailyPriceComparison {
-                    today_average_minor_units: Some(78900),
-                    yesterday_average_minor_units: Some(89900),
-                },
-            };
-            let json = serde_json::to_string(&result).unwrap();
-            assert!(json.contains("in_stock"));
-            assert!(json.contains("Back in Stock!"));
-        }
-
-        #[test]
-        fn test_with_none_notification() {
-            let check = AvailabilityCheckModel {
-                id: Uuid::new_v4(),
-                product_id: Uuid::new_v4(),
-                status: "out_of_stock".to_string(),
-                raw_availability: None,
-                error_message: None,
-                checked_at: chrono::Utc::now(),
-                price_minor_units: None,
-                price_currency: None,
-                raw_price: None,
-            };
-            let result = CheckResultWithNotification {
-                check,
-                notification: None,
-                daily_comparison: DailyPriceComparison::default(),
-            };
-            let json = serde_json::to_string(&result).unwrap();
-            assert!(json.contains("out_of_stock"));
-            assert!(json.contains("null") || !json.contains("notification"));
-        }
-    }
-
-    /// Tests for get_daily_price_comparison method
-    mod daily_price_comparison_tests {
-        use super::*;
-        use crate::repositories::AvailabilityCheckRepository;
-        use crate::test_utils::{create_test_product, setup_availability_db};
-
-        #[tokio::test]
-        async fn test_no_data() {
-            let conn = setup_availability_db().await;
-            let product_id = create_test_product(&conn, "https://example.com").await;
-
-            let comparison = AvailabilityService::get_daily_price_comparison(&conn, product_id)
-                .await
-                .unwrap();
-
-            assert_eq!(comparison.today_average_minor_units, None);
-            assert_eq!(comparison.yesterday_average_minor_units, None);
-        }
-
-        #[tokio::test]
-        async fn test_today_only() {
-            let conn = setup_availability_db().await;
-            let product_id = create_test_product(&conn, "https://example.com").await;
-
-            // Check 6 hours ago — within "today" (last 24h)
-            let check_time = chrono::Utc::now() - chrono::Duration::hours(6);
-            AvailabilityCheckRepository::create_with_timestamp(
-                &conn,
-                product_id,
-                Some(15000),
-                check_time,
-            )
-            .await;
-
-            let comparison = AvailabilityService::get_daily_price_comparison(&conn, product_id)
-                .await
-                .unwrap();
-
-            assert_eq!(comparison.today_average_minor_units, Some(15000));
-            assert_eq!(comparison.yesterday_average_minor_units, None);
-        }
-
-        #[tokio::test]
-        async fn test_today_and_yesterday() {
-            let conn = setup_availability_db().await;
-            let product_id = create_test_product(&conn, "https://example.com").await;
-
-            // Today: 6 hours ago
-            AvailabilityCheckRepository::create_with_timestamp(
-                &conn,
-                product_id,
-                Some(15000),
-                chrono::Utc::now() - chrono::Duration::hours(6),
-            )
-            .await;
-
-            // Yesterday: 30 hours ago (within 24–48h window)
-            AvailabilityCheckRepository::create_with_timestamp(
-                &conn,
-                product_id,
-                Some(20000),
-                chrono::Utc::now() - chrono::Duration::hours(30),
-            )
-            .await;
-
-            let comparison = AvailabilityService::get_daily_price_comparison(&conn, product_id)
-                .await
-                .unwrap();
-
-            assert_eq!(comparison.today_average_minor_units, Some(15000));
-            assert_eq!(comparison.yesterday_average_minor_units, Some(20000));
-        }
-
-        #[tokio::test]
-        async fn test_yesterday_only() {
-            let conn = setup_availability_db().await;
-            let product_id = create_test_product(&conn, "https://example.com").await;
-
-            // 30 hours ago — within "yesterday" (24–48h window)
-            AvailabilityCheckRepository::create_with_timestamp(
-                &conn,
-                product_id,
-                Some(20000),
-                chrono::Utc::now() - chrono::Duration::hours(30),
-            )
-            .await;
-
-            let comparison = AvailabilityService::get_daily_price_comparison(&conn, product_id)
-                .await
-                .unwrap();
-
-            assert_eq!(comparison.today_average_minor_units, None);
-            assert_eq!(comparison.yesterday_average_minor_units, Some(20000));
-        }
-
-        #[tokio::test]
-        async fn test_old_data_excluded() {
-            let conn = setup_availability_db().await;
-            let product_id = create_test_product(&conn, "https://example.com").await;
-
-            // 72 hours ago — beyond both windows
-            AvailabilityCheckRepository::create_with_timestamp(
-                &conn,
-                product_id,
-                Some(20000),
-                chrono::Utc::now() - chrono::Duration::hours(72),
-            )
-            .await;
-
-            let comparison = AvailabilityService::get_daily_price_comparison(&conn, product_id)
-                .await
-                .unwrap();
-
-            assert_eq!(comparison.today_average_minor_units, None);
-            assert_eq!(comparison.yesterday_average_minor_units, None);
         }
     }
